+++ app/client/cypress/e2e/Regression/ClientSide/Anvil/Widgets/AnvilSwitchWidgetSnapshot_spec.ts
@@ -5,7 +5,7 @@ import {
 } from "../../../../../support/Objects/ObjectsCore";
 
 // TODO: Enable when issue(github.com/appsmithorg/appsmith/issues/36419) is solved.
-describe(
+describe.skip(
   `${ANVIL_EDITOR_TEST}: Anvil tests for Switch Widget`,
   { tags: ["@tag.Anvil", "@tag.Visual"] },
   () => {
