+++ app/client/packages/design-system/widgets/src/components/Calendar/src/CalendarYearDropdown.tsx
@@ -1,25 +1,12 @@
-import { ListBoxItem, Select } from "@appsmith/wds";
-import { useDateFormatter } from "@react-aria/i18n";
-import type { CalendarState } from "@react-stately/calendar";
 import React from "react";
 import type { Key } from "react";
+import { ListBoxItem, Select } from "@appsmith/wds";
+import type { CalendarState } from "@react-stately/calendar";
+
+import { useYearOptions } from "../utils/calendar";
 
 export function CalendarYearDropdown({ state }: { state: CalendarState }) {
-  const years: { value: CalendarState["focusedDate"]; formatted: string }[] =
-    [];
-  const formatter = useDateFormatter({
-    year: "numeric",
-    timeZone: state.timeZone,
-  });
-
-  for (let i = -20; i <= 20; i++) {
-    const date = state.focusedDate.add({ years: i });
-
-    years.push({
-      value: date,
-      formatted: formatter.format(date.toDate(state.timeZone)),
-    });
-  }
+  const years = useYearOptions(state);
 
   const onChange = (value: Key | null) => {
     const index = Number(value);
